Improving UX of messages

+++ languages/languageDesign/messages/source_gen/jetbrains/mps/lang/messages/editor/MessageExpression_Actions.java
@@ -8,8 +8,12 @@ import jetbrains.mps.openapi.editor.EditorContext;
 import jetbrains.mps.lang.smodel.generator.smodelAdapter.SNodeOperations;
 import jetbrains.mps.smodel.adapter.structure.MetaAdapterFactory;
 import jetbrains.mps.lang.smodel.generator.smodelAdapter.SConceptOperations;
+import jetbrains.mps.lang.smodel.generator.smodelAdapter.SPropertyOperations;
 import jetbrains.mps.internal.collections.runtime.ListSequence;
 import jetbrains.mps.lang.smodel.generator.smodelAdapter.SLinkOperations;
+import jetbrains.mps.editor.runtime.deletionApprover.DeletionApproverUtil;
+import jetbrains.mps.editor.runtime.selection.SelectionUtil;
+import jetbrains.mps.openapi.editor.selection.SelectionManager;
 import jetbrains.mps.openapi.editor.cells.EditorCell;
 import jetbrains.mps.openapi.editor.cells.CellAction;
 import jetbrains.mps.openapi.editor.cells.CellActionType;
@@ -25,6 +29,23 @@ public class MessageExpression_Actions {
       public void execute_internal(EditorContext editorContext, SNode node) {
         if (SNodeOperations.isInstanceOf(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"))) {
           SNode newMessage = SConceptOperations.createNewNode(MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"));
+          SPropertyOperations.assign(newMessage, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message"), "");
+          new LiteralMessageSplitHelper(editorContext, SNodeOperations.cast(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"))).insertMessageExpr(newMessage);
+          editorContext.selectWRTFocusPolicy(newMessage);
+        }
+      }
+
+    };
+  }
+  /*package*/ static AbstractCellAction createAction_INSERT_BEFORE(final SNode node) {
+    return new AbstractCellAction() {
+      public void execute(EditorContext editorContext) {
+        this.execute_internal(editorContext, node);
+      }
+      public void execute_internal(EditorContext editorContext, SNode node) {
+        if (SNodeOperations.isInstanceOf(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"))) {
+          SNode newMessage = SConceptOperations.createNewNode(MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"));
+          SPropertyOperations.assign(newMessage, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message"), "");
           new LiteralMessageSplitHelper(editorContext, SNodeOperations.cast(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"))).insertMessageExpr(newMessage);
           editorContext.selectWRTFocusPolicy(newMessage);
         }
@@ -41,7 +62,76 @@ public class MessageExpression_Actions {
         if (SNodeOperations.hasRole(node, MetaAdapterFactory.getContainmentLink(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e40455fL, 0x48f860fc0e404561L, "part")) && ListSequence.fromList(SLinkOperations.getChildren(SNodeOperations.cast(SNodeOperations.getParent(node), MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e40455fL, "jetbrains.mps.lang.messages.structure.CombinedMessageExpression")), MetaAdapterFactory.getContainmentLink(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e40455fL, 0x48f860fc0e404561L, "part"))).count() == 1) {
           SNodeOperations.deleteNode(SNodeOperations.getParent(node));
         } else {
-          SNodeOperations.deleteNode(node);
+          if (SNodeOperations.isInstanceOf(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e41e4a2L, "jetbrains.mps.lang.messages.structure.MacroMessageExpression"))) {
+            if (DeletionApproverUtil.approve(editorContext, node)) {
+              return;
+            }
+            SNodeOperations.deleteNode(node);
+          } else if (SNodeOperations.isInstanceOf(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"))) {
+            SNode lit = SNodeOperations.cast(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"));
+            if (isEmptyString(SPropertyOperations.getString(lit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message")))) {
+              SNode nextLit = SNodeOperations.as(SNodeOperations.getNextSibling(lit), MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"));
+              SNodeOperations.deleteNode(lit);
+              if (nextLit != null) {
+                SelectionUtil.selectLabelCellAnSetCaret(editorContext, nextLit, SelectionManager.FIRST_CELL, 0);
+              }
+            } else {
+              SNode nextLit = SNodeOperations.as(SNodeOperations.getNextSibling(lit), MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"));
+              if (nextLit != null) {
+                if (isNotEmptyString(SPropertyOperations.getString(nextLit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message")))) {
+                  SPropertyOperations.plusAssignStringProp(lit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message"), SPropertyOperations.getString(nextLit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message")).substring(1));
+                }
+                SNodeOperations.deleteNode(nextLit);
+              } else {
+                if (DeletionApproverUtil.approve(editorContext, SNodeOperations.getParent(node))) {
+                  return;
+                }
+                SNodeOperations.deleteNode(SNodeOperations.getParent(node));
+              }
+            }
+          }
+        }
+      }
+
+    };
+  }
+  /*package*/ static AbstractCellAction createAction_BACKSPACE(final SNode node) {
+    return new AbstractCellAction() {
+      public void execute(EditorContext editorContext) {
+        this.execute_internal(editorContext, node);
+      }
+      public void execute_internal(EditorContext editorContext, SNode node) {
+        if (SNodeOperations.hasRole(node, MetaAdapterFactory.getContainmentLink(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e40455fL, 0x48f860fc0e404561L, "part")) && ListSequence.fromList(SLinkOperations.getChildren(SNodeOperations.cast(SNodeOperations.getParent(node), MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e40455fL, "jetbrains.mps.lang.messages.structure.CombinedMessageExpression")), MetaAdapterFactory.getContainmentLink(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e40455fL, 0x48f860fc0e404561L, "part"))).count() == 1) {
+          SNodeOperations.deleteNode(SNodeOperations.getParent(node));
+        } else {
+          if (SNodeOperations.isInstanceOf(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e41e4a2L, "jetbrains.mps.lang.messages.structure.MacroMessageExpression"))) {
+            if (DeletionApproverUtil.approve(editorContext, node)) {
+              return;
+            }
+            SNodeOperations.deleteNode(node);
+          } else if (SNodeOperations.isInstanceOf(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"))) {
+            SNode lit = SNodeOperations.cast(node, MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"));
+            if (isEmptyString(SPropertyOperations.getString(lit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message")))) {
+              SNode prevLit = SNodeOperations.as(SNodeOperations.getPrevSibling(lit), MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"));
+              SNodeOperations.deleteNode(lit);
+              if (prevLit != null) {
+                SelectionUtil.selectLabelCellAnSetCaret(editorContext, prevLit, SelectionManager.LAST_CELL, -1);
+              }
+            } else {
+              SNode prevLit = SNodeOperations.as(SNodeOperations.getPrevSibling(lit), MetaAdapterFactory.getConcept(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, "jetbrains.mps.lang.messages.structure.LiteralMessageExpression"));
+              if (prevLit != null) {
+                if (isNotEmptyString(SPropertyOperations.getString(prevLit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message")))) {
+                  SPropertyOperations.assign(lit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message"), SPropertyOperations.getString(prevLit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message")).substring(0, SPropertyOperations.getString(prevLit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message")).length() - 1) + SPropertyOperations.getString(lit, MetaAdapterFactory.getProperty(0xad93155d79b24759L, 0xb10c55123e763903L, 0x48f860fc0e362dc5L, 0x48f860fc0e362dc6L, "message")));
+                }
+                SNodeOperations.deleteNode(prevLit);
+              } else {
+                if (DeletionApproverUtil.approve(editorContext, SNodeOperations.getParent(node))) {
+                  return;
+                }
+                SNodeOperations.deleteNode(SNodeOperations.getParent(node));
+              }
+            }
+          }
         }
       }
 
@@ -70,7 +160,9 @@ public class MessageExpression_Actions {
 
     // set cell actions defined directly in this action map 
     editorCell.setAction(CellActionType.INSERT, createAction_INSERT(node));
+    editorCell.setAction(CellActionType.INSERT_BEFORE, createAction_INSERT_BEFORE(node));
     editorCell.setAction(CellActionType.DELETE, createAction_DELETE(node));
+    editorCell.setAction(CellActionType.BACKSPACE, createAction_BACKSPACE(node));
 
   }
 
@@ -82,8 +174,20 @@ public class MessageExpression_Actions {
     if (Objects.equals(actionType, CellActionType.INSERT)) {
       editorCell.setAction(actionType, createAction_INSERT(node));
     }
+    if (Objects.equals(actionType, CellActionType.INSERT_BEFORE)) {
+      editorCell.setAction(actionType, createAction_INSERT_BEFORE(node));
+    }
     if (Objects.equals(actionType, CellActionType.DELETE)) {
       editorCell.setAction(actionType, createAction_DELETE(node));
     }
+    if (Objects.equals(actionType, CellActionType.BACKSPACE)) {
+      editorCell.setAction(actionType, createAction_BACKSPACE(node));
+    }
+  }
+  private static boolean isEmptyString(String str) {
+    return str == null || str.length() == 0;
+  }
+  private static boolean isNotEmptyString(String str) {
+    return str != null && str.length() > 0;
   }
 }
